buffer

+++ RapidOCRSharpOnnx/Utils/FixedBuffer.cs
@@ -9,11 +9,11 @@ namespace RapidOCRSharpOnnx.Utils
 {
     public unsafe sealed class FixedBuffer : IDisposable
     {
-        public long Length { get; }
+        public int Length { get; }
         public IntPtr Address => (IntPtr)Pointer;
         public float* Pointer { get; private set; }
 
-        public FixedBuffer(long length)
+        public FixedBuffer(int length)
         {
             Length = length;
             Pointer = (float*)NativeMemory.Alloc((nuint)length, sizeof(float));
